started spliting views in separate files
